Now uses SevSeg library for support of multiple types of displays.
SevSeg library supports both common anode and cathode. Please test. From
here, I plan to make many smaller firmware changes (that do not related
to how segments are actually displayed).

I need to get the products with a ATmega type backpack (Serial7Segment
and OpenSegment) to this library so that we can use similar code, but
with different pin definitions.