Added RX pin factory reset feature. Moved away from hardware UART setup.
Many changes but the major one was adding the ability to reset the UART
to 9600 when the RX is pulled low for a certain amount of time. This
affected boot up, EEPROM locations, and the way that the UART is
controlled. We now initialize the UART using standard Arduino calls
instead of hardware registers. This commit works well with the
BasicSerial example code. Even though there are quite a few code
changes, the test code (BasicSerial) leads me to believe it works bug
free.