fixed some comments, re-worked pin labels in sch.
Fixed comments in System_Functions.ino which mapped Arduino pins to
display pins
Some labels on the schematic did not match the display pin they
connected to. Renamed those nets to make more sense.
Added a pdf of current schematic, not sure if this'll be sustainable.
Trying it out.
Defaulted display type to S7S in the firmware.