updates firmware to support SegSev 3.7.0
This old firmware is designed to be used with a separately maintained
and released fork of SegSev and the APIs now appear to differ.
These updates address this by updating various calls to the new API
and making the necessary refresh calls.

This change is probably useful for others since folks using library
managers will likely end up with the other version.

Note that this has only been tested on the S7S board layout. Open segment
and 7 segment shield were updated but are untested.